+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Passage en vue matérialisée de la vue `adresse.v_certificat`.
+* Ajout d'un algorithme dans la boite à outil Processing pour provoquer le rafraîchissement de la vue ci-dessus.
+
 ## 1.0.0 - 2021-11-10
 
 * Amélioration de la fonction `get_parcelle` pour qu'elle retourne une valeur unique